requirements.txt
dotenv==0.9.9
dspy==2.6.10
huggingface-hub==0.29.2
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
litellm==1.63.0
openai==1.65.2
pinecone==6.0.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
sentence-transformers
mlflow
playwright
pytest
fastapi
uvicorn
slowapi
pydantic-settings
uvloop
ragas
